pyproject.toml
[tool.poetry]
name = "VirtuGhan"
version = "0.3.4"
description = "Virtual Computation Cube"
authors = ["kshitijrajsharma <skshitizraj@gmail.com>"]
readme = "README.md"
packages = [
    { include = "vcube", from = "src" }
]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.10.0"
rio-tiler = "^7.2.2"
fastapi = "^0.115.6"
pillow = "^11.0.0"
tqdm = "^4.67.1"
mercantile = "^1.2.1"
numpy = "^2.2.1"
shapely = "^2.0.6"
httpx = "^0.28.1"
aiocache = "^0.12.3"
jinja2 = "^3.1.5"
uvicorn = "^0.34.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
commitizen = "^4.1.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "poetry"
update_changelog_on_bump = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
